Publish Version 0.0.1-r1
The Version 0.0.1-r1, a very early demo version with no fancy effects.
It can't be downloaded by others, because all configuration is encoded
in the code and can not adjust to different environments.

It will be usable when imported with yaml config file.

The interface needs to be modified and the shader also. Document model
needs to be refined and so as the rfc text doc management.